remove universal recipient registry
