+++ subgraph/config/arbitrum-rinkeby.json
@@ -2,6 +2,5 @@
   "network": "arbitrum-rinkeby",
   "address": "0xC032e80a413Be959d9a9B6e5CadE53c870074d37",
   "factoryStartBlock": 4806990,
-  "recipientRegistryStartBlock": 4806990,
-  "universalRecipientRegistryStartBlock": 4806990
+  "recipientRegistryStartBlock": 4806990
 }
