+++ subgraph/config/arbitrum.json
@@ -2,6 +2,5 @@
   "network": "arbitrum-one",
   "address": "0x2e89494a8fE02891511a43f7877b726787E0C160",
   "factoryStartBlock": 3461582,
-  "recipientRegistryStartBlock": 3461582,
-  "universalRecipientRegistryStartBlock": 3461582
+  "recipientRegistryStartBlock": 3461582
 }
