+++ subgraph/src/recipientRegistry/RecipientRegistryType.ts
@@ -1,21 +1,18 @@
 import { Address, TypedMap, log } from '@graphprotocol/graph-ts'
 import { OptimisticRecipientRegistry as OptimisticRecipientRegistryContract } from '../../generated/FundingRoundFactory/OptimisticRecipientRegistry'
 import { KlerosRecipientRegistry as KlerosRecipientRegistryContract } from '../../generated/FundingRoundFactory/KlerosRecipientRegistry'
-import { UniversalRecipientRegistry as UniversalRecipientRegistryContract } from '../../generated/FundingRoundFactory/UniversalRecipientRegistry'
 
 export enum RecipientRegistryType {
   Unknown,
   Simple,
   Kleros,
   Optimistic,
-  Universal,
 }
 
 let registryTypeMap = new TypedMap<string, RecipientRegistryType>()
 registryTypeMap.set('simple', RecipientRegistryType.Simple)
 registryTypeMap.set('klerso', RecipientRegistryType.Kleros)
 registryTypeMap.set('optimistic', RecipientRegistryType.Optimistic)
-registryTypeMap.set('universal', RecipientRegistryType.Universal)
 
 /**
  * Determine the type of the registry given the registry address
@@ -31,19 +28,6 @@ registryTypeMap.set('universal', RecipientRegistryType.Universal)
 export function getRecipientRegistryType(
   registryAddress: Address
 ): RecipientRegistryType {
-  let universalRegistry =
-    UniversalRecipientRegistryContract.bind(registryAddress)
-  let registryType = universalRegistry.try_registryType()
-  if (!registryType.reverted) {
-    let entry = registryTypeMap.getEntry(registryType.value)
-    if (entry) {
-      return entry.value
-    } else {
-      log.error('Error: unknown registry type {}', [registryType.value])
-      return RecipientRegistryType.Unknown
-    }
-  }
-
   let klerosRegistry = KlerosRecipientRegistryContract.bind(registryAddress)
   let tcr = klerosRegistry.try_tcr()
   if (!tcr.reverted) {
