+++ vue-app/.env.xdai
@@ -19,7 +19,7 @@ VUE_APP_USER_REGISTRY_TYPE=brightid
 # Learn more about BrightID and context in /docs/brightid.md
 VUE_APP_BRIGHTID_CONTEXT=clr.fund
 
-# Supported values: simple, optimistic, kleros, universal
+# Supported values: simple, optimistic, kleros
 VUE_APP_RECIPIENT_REGISTRY_TYPE=optimistic
 
 VUE_APP_RECIPIENT_REGISTRY_POLICY=QmeygKjvrpidJeFHv6ywjUrj718nwtFQgCCPPR4r5nL87R
