+++ vue-app/src/api/abi.ts
@@ -7,7 +7,6 @@ import { abi as UserRegistry } from '../../../contracts/build/contracts/contract
 import { abi as BrightIdUserRegistry } from '../../../contracts/build/contracts/contracts/userRegistry/BrightIdUserRegistry.sol/BrightIdUserRegistry.json'
 import { abi as SimpleRecipientRegistry } from '../../../contracts/build/contracts/contracts/recipientRegistry/SimpleRecipientRegistry.sol/SimpleRecipientRegistry.json'
 import { abi as OptimisticRecipientRegistry } from '../../../contracts/build/contracts/contracts/recipientRegistry/OptimisticRecipientRegistry.sol/OptimisticRecipientRegistry.json'
-import { abi as UniversalRecipientRegistry } from '../../../contracts/build/contracts/contracts/recipientRegistry/UniversalRecipientRegistry.sol/UniversalRecipientRegistry.json'
 import { abi as KlerosGTCR } from '../../../contracts/build/contracts/contracts/recipientRegistry/IKlerosGTCR.sol/IKlerosGTCR.json'
 import { abi as KlerosGTCRAdapter } from '../../../contracts/build/contracts/contracts/recipientRegistry/KlerosGTCRAdapter.sol/KlerosGTCRAdapter.json'
 import { abi as BaseRecipientRegistry } from '../../../contracts/build/contracts/contracts/recipientRegistry/BaseRecipientRegistry.sol/BaseRecipientRegistry.json'
@@ -22,7 +21,6 @@ export {
   BrightIdUserRegistry,
   SimpleRecipientRegistry,
   OptimisticRecipientRegistry,
-  UniversalRecipientRegistry,
   KlerosGTCR,
   KlerosGTCRAdapter,
   BaseRecipientRegistry,
