+++ vue-app/src/views/JoinOptimistic.vue
@@ -15,698 +15,55 @@
       />
       <div class="title-area">
         <h1>Join the round</h1>
-        <div v-if="currentStep === 5">
-          <div class="toggle-tabs-desktop">
-            <p
-              class="tab"
-              id="review"
-              :class="showSummaryPreview ? 'inactive-tab' : 'active-tab'"
-              @click="handleToggleTab"
-            >
-              Review info
-            </p>
-            <p
-              class="tab"
-              id="preview"
-              :class="showSummaryPreview ? 'active-tab' : 'inactive-tab'"
-              @click="handleToggleTab"
-            >
-              Preview project
-            </p>
-          </div>
-        </div>
       </div>
       <div class="cancel-area desktop">
         <links class="cancel-link" to="/join"> Cancel </links>
       </div>
-      <div class="form-area">
+      <loader v-if="loading" />
+      <div v-if="!loading" class="form-area">
         <div class="application">
-          <div v-if="currentStep === 0">
-            <h2 class="step-title">About the project</h2>
-            <div class="inputs">
-              <div class="form-background">
-                <label for="project-name" class="input-label"
-                  >Project name</label
-                >
-                <input
-                  id="project-name"
-                  type="text"
-                  placeholder="ex: clr.fund"
-                  v-model="$v.form.project.name.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.project.name.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.project.name.$error,
-                  }"
-                >
-                  Your project needs a name
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="project-tagline" class="input-label">Tagline</label>
-                <p class="input-description">
-                  Describe your project in a sentence. Max characters: 140
-                </p>
-                <textarea
-                  id="project-tagline"
-                  placeholder="ex: A quadratic funding protocol"
-                  v-model="$v.form.project.tagline.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.project.tagline.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden:
-                      !$v.form.project.tagline.$error ||
-                      $v.form.project.tagline.maxLength,
-                  }"
-                >
-                  This tagline is too long. Be brief for potential contributors
-                </p>
-                <p
-                  :class="{
-                    error: true,
-                    hidden:
-                      !$v.form.project.tagline.$error ||
-                      !$v.form.project.tagline.maxLength,
-                  }"
-                >
-                  Your project needs a tagline
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="project-description" class="input-label">
-                  Description
-                  <p class="input-description">Markdown supported.</p>
-                </label>
-                <textarea
-                  id="project-description"
-                  placeholder="ex: CLR.fund is a quadratic funding protocol that aims to make it as easy as possible to set up, manage, and participate in quadratic funding rounds..."
-                  v-model="$v.form.project.description.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.project.description.$error,
-                  }"
-                />
-                <p v-if="form.project.description" class="input-label pt-1">
-                  Preview:
-                </p>
-                <markdown :raw="form.project.description" />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.project.description.$error,
-                  }"
-                >
-                  Your project needs a description. What are you raising money
-                  for?
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="project-category" class="input-label"
-                  >Category
-                  <p class="input-description">Choose the best fit.</p>
-                </label>
-                <form class="radio-row" id="category-radio">
-                  <input
-                    id="category-content"
-                    type="radio"
-                    name="project-category"
-                    value="Content"
-                    v-model="$v.form.project.category.$model"
-                    :class="{
-                      input: true,
-                      invalid: $v.form.project.category.$error,
-                    }"
-                  />
-                  <label for="category-content" class="radio-btn"
-                    >Content</label
-                  >
-                  <input
-                    id="research"
-                    type="radio"
-                    name="project-category"
-                    value="Research"
-                    v-model="$v.form.project.category.$model"
-                    :class="{
-                      input: true,
-                      invalid: $v.form.project.category.$error,
-                    }"
-                  />
-                  <label for="research" class="radio-btn">Research</label>
-                  <input
-                    id="tooling"
-                    type="radio"
-                    name="project-category"
-                    value="Tooling"
-                    v-model="$v.form.project.category.$model"
-                    :class="{
-                      input: true,
-                      invalid: $v.form.project.category.$error,
-                    }"
-                  />
-                  <label for="tooling" class="radio-btn">Tooling</label>
-                  <input
-                    id="data"
-                    type="radio"
-                    name="project-category"
-                    value="Data"
-                    v-model="$v.form.project.category.$model"
-                    :class="{
-                      input: true,
-                      invalid: $v.form.project.category.$error,
-                    }"
-                  />
-                  <label for="data" class="radio-btn">Data</label>
-                </form>
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.project.category.$error,
-                  }"
-                >
-                  You need to choose a category
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="project-problem-space" class="input-label"
-                  >Problem space</label
-                >
-                <p class="input-description">
-                  Explain the problems you're trying to solve. Markdown
-                  supported.
-                </p>
-                <textarea
-                  id="project-problem-space"
-                  placeholder="ex: there is no way to spin up a quadratic funding round. Right now, you have to collaborate with GitCoin Grants which isn’t a scalable or sustainable model."
-                  v-model="$v.form.project.problemSpace.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.project.problemSpace.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.project.problemSpace.$error,
-                  }"
-                >
-                  Explain the problem your project solves
-                </p>
-                <p v-if="form.project.description" class="input-label pt-1">
-                  Preview:
-                </p>
-                <markdown :raw="form.project.problemSpace" />
-              </div>
-            </div>
+          <div v-if="steps[currentStep] === 'metadata'">
+            <metadata-list
+              :onClick="handleMetadataSelected"
+              :excludeRecipients="true"
+            ></metadata-list>
           </div>
-          <div v-if="currentStep === 1">
-            <h2 class="step-title">Donation details</h2>
-            <div class="inputs">
-              <div class="form-background">
-                <label for="fund-address" class="input-label"
-                  >Ethereum address</label
-                >
-                <p class="input-description">
-                  The destination address for donations, which you'll use to
-                  claim funds. This doesn't have to be the same address as the
-                  one you use to send your application transaction.
-                </p>
-                <input
-                  id="fund-address"
-                  placeholder="example: 0x123..."
-                  v-model.lazy="$v.form.fund.addressName.$model"
-                  @blur="checkEns"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.fund.addressName.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.fund.addressName.$error,
-                  }"
-                >
-                  Enter a valid ENS or Ethereum 0x address
-                </p>
-                <!-- TODO: only validate after user removes focus on input -->
-              </div>
-              <div class="form-background">
-                <label for="fund-plans" class="input-label"
-                  >How will you spend your funding?</label
-                >
-                <p class="input-description">
-                  Potential contributors might convert based on your specific
-                  funding plans. Markdown supported.
-                </p>
-                <textarea
-                  id="fund-plans"
-                  placeholder="ex: on our roadmap..."
-                  v-model="$v.form.fund.plans.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.fund.plans.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.fund.plans.$error,
-                  }"
-                >
-                  Let potential contributors know what plans you have for their
-                  donations.
-                </p>
-                <p v-if="form.fund.plans" class="input-label pt-1">Preview:</p>
-                <markdown :raw="form.fund.plans" />
-              </div>
-            </div>
+          <div v-if="steps[currentStep] === 'summary'" id="summary">
+            <metadata-viewer :metadata="metadata"></metadata-viewer>
           </div>
-          <div v-if="currentStep === 2">
-            <h2 class="step-title">Team details</h2>
-            <p>Tell us about the folks behind your project.</p>
+          <div v-if="steps[currentStep] === 'email'">
             <div class="inputs">
-              <div v-if="isEmailRequired" class="form-background">
-                <label for="team-email" class="input-label">
-                  Contact email
-                </label>
-                <p class="input-description">
-                  For important updates about your project and the funding
-                  round.
-                </p>
-                <input
-                  id="team-email"
-                  type="email"
-                  placeholder="example: doge@goodboi.com"
-                  v-model.lazy="$v.form.team.email.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.team.email.$error,
-                  }"
-                />
-                <p class="input-notice">
-                  We won't display this publicly or add it to the on-chain
-                  registry.
-                </p>
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.team.email.$error,
-                  }"
-                >
-                  This doesn't look like an email.
-                </p>
+              <div class="heading">
+                <h2>Contact email</h2>
               </div>
-              <div class="form-background">
-                <label for="team-name" class="input-label"
-                  >Team name (optional)</label
-                >
-                <p class="input-description">If different to project name.</p>
-                <input
-                  id="team-name"
-                  placeholder="ex: clr.fund"
-                  v-model="$v.form.team.name.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.team.name.$error,
-                  }"
-                />
-              </div>
-              <div class="form-background">
-                <label for="team-desc" class="input-label"
-                  >Team description (optional)</label
-                >
-                <p class="input-description">
-                  If different to project description. Markdown supported.
-                </p>
-                <textarea
-                  id="team-desc"
-                  placeholder="ex: CLR.fund is a quadratic funding protocol that aims to make it as easy as possible to set up, manage, and participate in quadratic funding rounds..."
-                  v-model="$v.form.team.description.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.team.description.$error,
-                  }"
-                />
-                <p v-if="form.team.description" class="input-label pt-1">
-                  Preview:
-                </p>
-                <markdown :raw="form.team.description" />
-              </div>
-            </div>
-          </div>
-          <div v-if="currentStep === 3">
-            <h2 class="step-title">Links</h2>
-            <p>
-              Give contributors some links to check out to learn more about your
-              project. Provide at least one.
-            </p>
-            <div class="inputs">
-              <div class="form-background">
-                <label for="links-github" class="input-label">GitHub</label>
-                <input
-                  id="links-github"
-                  type="link"
-                  placeholder="example: https://github.com/ethereum/clrfund"
-                  v-model.lazy="$v.form.links.github.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.links.github.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.links.github.$error,
-                  }"
-                >
-                  This doesn't look like a valid URL
-                </p>
-                <!-- TODO: only validate after user removes focus on input -->
-              </div>
-              <div class="form-background">
-                <label for="links-radicle" class="input-label">Radicle</label>
-                <input
-                  id="links-radicle"
-                  type="link"
-                  placeholder="example: https://radicle.xyz/ethereum/clrfund"
-                  v-model.lazy="$v.form.links.radicle.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.links.radicle.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.links.radicle.$error,
-                  }"
-                >
-                  This doesn't look like a valid URL
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="links-website" class="input-label">Website</label>
-                <input
-                  id="links-website"
-                  type="link"
-                  placeholder="example: https://clr.fund"
-                  v-model.lazy="$v.form.links.website.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.links.website.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.links.website.$error,
-                  }"
-                >
-                  This doesn't look like a valid URL
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="links-twitter" class="input-label">Twitter</label>
-                <input
-                  id="links-twitter"
-                  type="link"
-                  placeholder="example: https://twitter.com/ethereum"
-                  v-model.lazy="$v.form.links.twitter.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.links.twitter.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.links.twitter.$error,
-                  }"
-                >
-                  This doesn't look like a valid URL
-                </p>
-              </div>
-              <div class="form-background">
-                <label for="links-discord" class="input-label">Discord</label>
-                <input
-                  id="links-discord"
-                  type="link"
-                  placeholder="ex: https://discord.gg/5Prub9zbGz"
-                  class="input"
-                  v-model.lazy="$v.form.links.discord.$model"
-                  :class="{
-                    input: true,
-                    invalid: $v.form.links.discord.$error,
-                  }"
-                />
-                <p
-                  :class="{
-                    error: true,
-                    hidden: !$v.form.links.discord.$error,
-                  }"
-                >
-                  This doesn't look like a valid URL
-                </p>
-              </div>
-            </div>
-          </div>
-          <div v-if="currentStep === 4">
-            <h2 class="step-title">Images</h2>
-            <p>
-              We'll upload your images to IPFS, a decentralized storage
-              platform.
-              <links to="https://ipfs.io/#how">More on IPFS</links>
-            </p>
-            <div class="inputs">
-              <div class="form-background">
-                <ipfs-image-upload
-                  label="Banner image"
-                  description="Recommended aspect ratio: 16x9 • Max file size: 512kB • JPG, PNG, or GIF"
-                  :onUpload="handleUpload"
-                  formProp="bannerHash"
-                  :hash="form.image.bannerHash"
-                />
-              </div>
-              <div class="form-background">
-                <ipfs-image-upload
-                  label="Thumbnail image"
-                  description="Recommended aspect ratio: 1x1 (square) • Max file size: 512kB • JPG, PNG, or GIF"
-                  :onUpload="handleUpload"
-                  formProp="thumbnailHash"
-                  :hash="form.image.thumbnailHash"
-                />
-              </div>
-            </div>
-          </div>
-          <div v-if="currentStep === 5" id="summary">
-            <project-profile
-              v-if="showSummaryPreview"
-              :project="projectInterface"
-              :previewMode="true"
-              class="project-details"
-            />
-            <div v-if="!showSummaryPreview">
-              <h2 class="step-title">Review your information</h2>
-              <warning
-                message="This information will be stored in a smart contract and cannot be edited, so please review carefully."
+              <p class="input-description">
+                For important updates about your project and the funding round.
+              </p>
+              <input
+                id="team-email"
+                type="email"
+                placeholder="example: doge@goodboi.com"
+                v-model.lazy="$v.metadata.email.$model"
+                :class="{
+                  input: true,
+                  invalid: $v.metadata.email.$error,
+                }"
               />
-              <div class="form-background">
-                <div class="summary-section-header">
-                  <h3 class="step-subtitle">About the project</h3>
-                  <links to="/join/project" class="edit-button"
-                    >Edit <img width="16px" src="@/assets/edit.svg"
-                  /></links>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Name</h4>
-                  <div class="data">{{ form.project.name }}</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Tagline</h4>
-                  <div class="data">{{ form.project.tagline }}</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Description</h4>
-                  <div class="data">{{ form.project.description }}</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Category</h4>
-                  <div class="data">{{ form.project.category }}</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Problem space</h4>
-                  <div class="data">{{ form.project.problemSpace }}</div>
-                </div>
-              </div>
-              <div class="form-background">
-                <div class="summary-section-header">
-                  <h3 class="step-subtitle">Funding details</h3>
-                  <links to="/join/fund" class="edit-button"
-                    >Edit <img width="16px" src="@/assets/edit.svg"
-                  /></links>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Ethereum address</h4>
-                  <div v-if="form.hasEns" class="data break-all">
-                    {{ form.fund.addressName }}
-                  </div>
-                  <div
-                    class="data"
-                    v-if="form.fund.addressName"
-                    title="Resolved ENS address"
-                  >
-                    <address-widget
-                      :address="form.fund.resolvedAddress"
-                      :chainId="chainId"
-                    />
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Funding plans</h4>
-                  <div class="data">{{ form.fund.plans }}</div>
-                </div>
-              </div>
-              <div class="form-background">
-                <div class="summary-section-header">
-                  <h3 class="step-subtitle">Team details</h3>
-                  <links to="/join/team" class="edit-button"
-                    >Edit <img width="16px" src="@/assets/edit.svg"
-                  /></links>
-                </div>
-                <div v-if="isEmailRequired" class="summary">
-                  <h4 class="read-only-title">Contact email</h4>
-                  <div class="data">{{ form.team.email }}</div>
-                  <div class="input-notice">
-                    This information won't be added to the smart contract.
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Team name</h4>
-                  <div class="data">{{ form.team.name }}</div>
-                  <div class="data" v-if="!form.team.name">Not provided</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Team description</h4>
-                  <div class="data">{{ form.team.description }}</div>
-                  <div class="data" v-if="!form.team.description">
-                    Not provided
-                  </div>
-                </div>
-              </div>
-              <div class="form-background">
-                <div class="summary-section-header">
-                  <h3 class="step-subtitle">Links</h3>
-                  <links to="/join/links" class="edit-button"
-                    >Edit <img width="16px" src="@/assets/edit.svg"
-                  /></links>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">GitHub</h4>
-                  <div class="data">
-                    {{ form.links.github }}
-                    <links
-                      v-if="form.links.github"
-                      :to="form.links.github"
-                      :hideArrow="true"
-                      ><img width="16px" src="@/assets/link.svg"
-                    /></links>
-                  </div>
-                  <div class="data" v-if="!form.links.github">Not provided</div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Twitter</h4>
-                  <div class="data">
-                    {{ form.links.twitter }}
-                    <links
-                      v-if="form.links.twitter"
-                      :to="form.links.twitter"
-                      :hideArrow="true"
-                      ><img width="16px" src="@/assets/link.svg"
-                    /></links>
-                  </div>
-                  <div class="data" v-if="!form.links.twitter">
-                    Not provided
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Website</h4>
-                  <div class="data" key="">
-                    {{ form.links.website }}
-                    <links
-                      v-if="form.links.website"
-                      :to="form.links.website"
-                      :hideArrow="true"
-                      ><img width="16px" src="@/assets/link.svg"
-                    /></links>
-                  </div>
-                  <div class="data" v-if="!form.links.website">
-                    Not provided
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Discord</h4>
-                  <div class="data">
-                    {{ form.links.discord }}
-                    <links
-                      v-if="form.links.discord"
-                      :to="form.links.discord"
-                      :hideArrow="true"
-                      ><img width="16px" src="@/assets/link.svg"
-                    /></links>
-                  </div>
-                  <div class="data" v-if="!form.links.discord">
-                    Not provided
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Radicle</h4>
-                  <div class="data">
-                    {{ form.links.radicle }}
-                    <links
-                      v-if="form.links.radicle"
-                      :to="form.links.radicle"
-                      :hideArrow="true"
-                      ><img width="16px" src="@/assets/link.svg"
-                    /></links>
-                  </div>
-                  <div class="data" v-if="!form.links.radicle">
-                    Not provided
-                  </div>
-                </div>
-              </div>
-              <div class="form-background">
-                <div class="summary-section-header">
-                  <h3 class="step-subtitle">Images</h3>
-                  <links to="/join/image" class="edit-button"
-                    >Edit <img width="16px" src="@/assets/edit.svg"
-                  /></links>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Banner</h4>
-                  <div class="data">
-                    <ipfs-copy-widget :hash="form.image.bannerHash" />
-                  </div>
-                </div>
-                <div class="summary">
-                  <h4 class="read-only-title">Thumbnail</h4>
-                  <div class="data">
-                    <ipfs-copy-widget :hash="form.image.thumbnailHash" />
-                  </div>
-                </div>
-              </div>
+              <p class="input-notice">
+                We won't display this publicly or add it to the on-chain
+                registry.
+              </p>
+              <p
+                :class="{
+                  error: true,
+                  hidden: !$v.metadata.email.$error,
+                }"
+              >
+                This doesn't look like an email.
+              </p>
             </div>
           </div>
-          <div v-if="currentStep === 6">
+          <div v-if="steps[currentStep] === 'submit'">
             <h2 class="step-title">Submit project</h2>
             <p>
               This is a blockchain transaction that will add your project
@@ -714,19 +71,20 @@
             </p>
             <div class="inputs">
               <recipient-submission-widget
-                cta="Submit project"
-                pending="Sending deposit..."
+                :isWaiting="isWaiting"
+                :txHash="txHash"
+                :txError="txError"
               />
             </div>
           </div>
         </div>
       </div>
     </div>
-    <div class="mobile nav-bar">
+    <div v-if="!loading" class="mobile nav-bar">
       <form-navigation
         :isStepValid="isStepValid(currentStep)"
         :steps="steps"
-        :finalStep="steps.length - 2"
+        :finalStep="steps.length - 1"
         :currentStep="currentStep"
         :callback="saveFormData"
         :handleStepNav="handleStepNav"
@@ -739,31 +97,27 @@
 <script lang="ts">
 import Component, { mixins } from 'vue-class-component'
 import { validationMixin } from 'vuelidate'
-import { required, maxLength, url, email } from 'vuelidate/lib/validators'
-import * as isIPFS from 'is-ipfs'
-import { isValidEthAddress, resolveEns } from '@/utils/accounts'
+import { required, email } from 'vuelidate/lib/validators'
 import LayoutSteps from '@/components/LayoutSteps.vue'
 import ProgressBar from '@/components/ProgressBar.vue'
 import FormNavigation from '@/components/FormNavigation.vue'
 import FormProgressWidget from '@/components/FormProgressWidget.vue'
-import IpfsImageUpload from '@/components/IpfsImageUpload.vue'
-import IpfsCopyWidget from '@/components/IpfsCopyWidget.vue'
-import AddressWidget from '@/components/AddressWidget.vue'
-import Loader from '@/components/Loader.vue'
-import Markdown from '@/components/Markdown.vue'
-import ProjectProfile from '@/components/ProjectProfile.vue'
 import RecipientSubmissionWidget from '@/components/RecipientSubmissionWidget.vue'
-import Warning from '@/components/Warning.vue'
+import MetadataList from '@/views/MetadataList.vue'
+import MetadataViewer from '@/views/MetadataViewer.vue'
 import Links from '@/components/Links.vue'
+import Loader from '@/components/Loader.vue'
 
-import { SET_RECIPIENT_DATA } from '@/store/mutation-types'
 import {
-  RecipientApplicationData,
-  formToProjectInterface,
-} from '@/api/recipient'
+  SET_RECIPIENT_DATA,
+  RESET_RECIPIENT_DATA,
+} from '@/store/mutation-types'
+
 import { Project } from '@/api/projects'
-import { chain } from '@/api/core'
-import { CHAIN_ID } from '@/plugins/Web3/constants/chains'
+import { Metadata } from '@/api/metadata'
+import { DateTime } from 'luxon'
+import { addRecipient } from '@/api/recipient-registry-optimistic'
+import { waitForTransaction } from '@/utils/contracts'
 
 @Component({
   components: {
@@ -771,200 +125,76 @@ import { CHAIN_ID } from '@/plugins/Web3/constants/chains'
     ProgressBar,
     FormNavigation,
     FormProgressWidget,
-    IpfsImageUpload,
-    IpfsCopyWidget,
-    AddressWidget,
-    Loader,
-    Markdown,
-    ProjectProfile,
     RecipientSubmissionWidget,
-    Warning,
+    MetadataList,
+    MetadataViewer,
     Links,
+    Loader,
   },
   validations: {
-    form: {
-      project: {
-        name: { required },
-        tagline: {
-          required,
-          maxLength: maxLength(140),
-        },
-        description: { required },
-        category: { required },
-        problemSpace: { required },
-      },
-      fund: {
-        addressName: {
-          required,
-          validEthAddress: isValidEthAddress,
-        },
-        resolvedAddress: {},
-        plans: { required },
-      },
-      team: {
-        name: {},
-        description: {},
-        email: {
-          email,
-          required: process.env.VUE_APP_GOOGLE_SPREADSHEET_ID
-            ? required
-            : () => true,
-        },
-      },
-      links: {
-        github: { url },
-        radicle: { url },
-        website: { url },
-        twitter: { url },
-        discord: { url },
-      },
-      image: {
-        bannerHash: {
-          required,
-          validIpfsHash: isIPFS.cid,
-        },
-        thumbnailHash: {
-          required,
-          validIpfsHash: isIPFS.cid,
-        },
+    metadata: {
+      email: {
+        email,
+        required,
       },
     },
   },
 })
 export default class JoinView extends mixins(validationMixin) {
-  form: RecipientApplicationData = {
-    project: {
-      name: '',
-      tagline: '',
-      description: '',
-      category: '',
-      problemSpace: '',
-    },
-    fund: {
-      receivingAddresses: [],
-      addressName: '',
-      resolvedAddress: '',
-      plans: '',
-    },
-    team: {
-      name: '',
-      description: '',
-      email: '',
-    },
-    links: {
-      github: '',
-      radicle: '',
-      website: '',
-      twitter: '',
-      discord: '',
-    },
-    image: {
-      bannerHash: '',
-      thumbnailHash: '',
-    },
-    furthestStep: 0,
-    hasEns: false,
-  }
   currentStep = 0
-  steps: string[] = []
-  stepNames: string[] = []
-  showSummaryPreview = false
+  steps = this.isEmailRequired
+    ? ['metadata', 'summary', 'email', 'submit']
+    : ['metadata', 'summary', 'submit']
+  stepNames = this.isEmailRequired
+    ? ['Select a project metadata', 'Review', 'Email', 'Submit']
+    : ['Select a project metadata', 'Review', 'Submit']
+  metadata = new Metadata({})
+  loading = true
+  isWaiting = false
+  txHash = ''
+  txError = ''
 
-  created() {
-    const steps = Object.keys(this.form)
-    steps.splice(steps.length - 1, 1, 'summary', 'submit')
-
-    const currentStep = steps.indexOf(this.$route.params.step)
-    const stepNames = [
-      'About the project',
-      'Donation details',
-      'Team details',
-      'Links',
-      'Images',
-      'Review',
-      'Submit',
-    ]
-    this.steps = steps
+  async created() {
+    const currentStep = this.steps.indexOf(this.$route.params.step)
     this.currentStep = currentStep
-    this.stepNames = stepNames
-    this.form = this.$store.state.recipient || this.form
 
-    // redirect to /join/ if step doesn't exist
+    // redirect to /join if step doesn't exist
     if (this.currentStep < 0) {
       this.$router.push({ name: 'join' })
     }
+
+    await this.loadMetadata()
+    this.loading = false
+
     // "Next" button restricts forward navigation via validation, and
     // eventually updates the `furthestStep` tracker when valid and clicked/tapped.
     // If URL step is ahead of furthest, navigate back to furthest
-    if (this.currentStep > this.form.furthestStep) {
+    if (this.currentStep > this.furthestStep) {
       this.$router.push({
         name: 'join-step',
-        params: { step: steps[this.form.furthestStep] },
+        params: { step: this.steps[this.furthestStep] },
       })
     }
   }
 
-  get isEmailRequired(): boolean {
-    return !!process.env.VUE_APP_GOOGLE_SPREADSHEET_ID
-  }
-
-  handleToggleTab(event): void {
-    const { id } = event.target
-    // Guard clause:
-    if (
-      (!this.showSummaryPreview && id === 'review') ||
-      (this.showSummaryPreview && id === 'preview')
-    )
-      return
-    this.showSummaryPreview = !this.showSummaryPreview
-  }
-
-  // Check that at least one link is not empty && no links are invalid
-  isLinkStepValid(): boolean {
-    let isValid = false
-    const links = Object.keys(this.form.links)
-    for (const link of links) {
-      const linkData = this.$v.form.links?.[link]
-      if (!linkData) return false
-      const isInvalid = linkData.$invalid
-      const isEmpty = linkData.$model.length === 0
-      if (isInvalid) {
-        return false
-      } else if (!isEmpty) {
-        isValid = true
-      }
-    }
-    return isValid
-  }
-
   isStepValid(step: number): boolean {
-    if (step === 3) {
-      return this.isLinkStepValid()
-    }
-    const stepName: string = this.steps[step]
-    return !this.$v.form[stepName]?.$invalid
+    return this.steps[step] === 'metadata'
+      ? this.hasMetadata
+      : this.hasMetadata && this.isMetadataValid
   }
 
   isStepUnlocked(step: number): boolean {
-    return step <= this.form.furthestStep
+    return step <= this.furthestStep
   }
 
-  saveFormData(updateFurthest?: boolean): void {
-    if (updateFurthest && this.currentStep + 1 > this.form.furthestStep) {
-      this.form.furthestStep = this.currentStep + 1
-    }
-    if (typeof this.currentStep !== 'number') return
+  saveFormData(): void {
     this.$store.commit(SET_RECIPIENT_DATA, {
-      updatedData: this.form,
-      step: this.steps[this.currentStep],
-      stepNumber: this.currentStep,
+      updatedData: this.metadata,
     })
   }
 
-  // Callback from IpfsImageUpload component
-  handleUpload(key, value) {
-    this.form.image[key] = value
-    this.saveFormData(false)
+  get isEmailRequired(): boolean {
+    return !!process.env.VUE_APP_GOOGLE_SPREADSHEET_ID
   }
 
   get isNavDisabled(): boolean {
@@ -974,41 +204,145 @@ export default class JoinView extends mixins(validationMixin) {
     )
   }
 
-  handleStepNav(step: number, updateFurthest?: boolean): void {
+  get hasMetadata(): boolean {
+    return !!this.metadata.id
+  }
+
+  get isMetadataValid(): boolean {
+    // TODO validate metadata
+    return true
+  }
+
+  get furthestStep(): number {
+    let furthest = 0
+    if (this.hasMetadata) {
+      furthest++
+      if (this.isMetadataValid) {
+        furthest++
+        if (
+          process.env.VUE_APP_GOOGLE_SPREADSHEET_ID &&
+          !this.$v.metadata.email?.$invalid
+        ) {
+          furthest++
+        }
+      }
+    }
+    return furthest
+  }
+
+  handleMetadataSelected(metadata: Metadata): void {
+    this.metadata = metadata
+    const id = this.metadata.id || ''
+    this.$router.push({
+      name: 'join-step',
+      params: { step: 'summary', id },
+    })
+  }
+
+  async handleStepNav(step: number): Promise<void> {
     // If isNavDisabled => disable quick-links
     if (this.isNavDisabled) return
     // Save form data
-    this.saveFormData(updateFurthest)
-    // Navigate
-    if (this.isStepUnlocked(step)) {
-      this.$router.push({
-        name: 'join-step',
-        params: {
-          step: this.steps[step],
-        },
-      })
+    this.saveFormData()
+
+    if (step >= this.steps.length) {
+      await this.addRecipient()
+    } else {
+      // Navigate
+      if (this.isStepUnlocked(step)) {
+        const id = this.metadata.id || ''
+        this.$router.push({
+          name: 'join-step',
+          params: {
+            step: this.steps[step],
+            id,
+          },
+        })
+      }
+    }
+  }
+
+  async loadMetadata(): Promise<void> {
+    const id = this.$route.params.id
+
+    if (id) {
+      if (id === this.$store.state.recipient?.id) {
+        this.metadata = new Metadata(this.$store.state.recipient)
+      } else {
+        try {
+          const metadata = await Metadata.get(id)
+          if (metadata) {
+            this.metadata = metadata
+            this.saveFormData()
+          }
+        } catch (err) {
+          // TODO display the error
+        }
+      }
     }
   }
 
   get projectInterface(): Project {
-    return formToProjectInterface(this.form)
+    return this.metadata.toProject()
   }
 
-  get furthestStep() {
-    return this.form.furthestStep
-  }
-
-  async checkEns(): Promise<void> {
-    const { addressName } = this.form?.fund
-    if (addressName) {
-      const res: string | null = await resolveEns(addressName)
-      this.form.hasEns = !!res
-      this.form.fund.resolvedAddress = res ? res : addressName
+  private async addRecipient() {
+    const {
+      currentRound,
+      currentUser,
+      recipient,
+      recipientRegistryAddress,
+      recipientRegistryInfo,
+    } = this.$store.state
+    this.isWaiting = true
+    // Reset errors when submitting
+    this.txError = ''
+    if (
+      recipientRegistryAddress &&
+      recipient &&
+      recipientRegistryInfo &&
+      currentUser
+    ) {
+      try {
+        if (currentRound && DateTime.now() >= currentRound.votingDeadline) {
+          this.$router.push({
+            name: 'join',
+          })
+          throw { message: 'round over' }
+        }
+        await waitForTransaction(
+          addRecipient(
+            recipientRegistryAddress,
+            new Metadata(recipient).toFormData(),
+            recipientRegistryInfo.deposit,
+            currentUser.walletProvider.getSigner()
+          ),
+          (hash) => (this.txHash = hash)
+        )
+        // Send application data to a Google Spreadsheet
+        if (process.env.VUE_APP_GOOGLE_SPREADSHEET_ID) {
+          await fetch('/.netlify/functions/recipient', {
+            method: 'POST',
+            headers: {
+              'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(recipient),
+          })
+        }
+        this.$store.commit(RESET_RECIPIENT_DATA)
+      } catch (error) {
+        this.isWaiting = false
+        this.txError = (error as any).message
+        return
+      }
+      this.isWaiting = false
+      this.$router.push({
+        name: 'project-added',
+        params: {
+          hash: this.txHash,
+        },
+      })
     }
-  }
-
-  get chainId(): number | undefined {
-    return CHAIN_ID[chain.name]
   }
 }
 </script>
@@ -1103,17 +437,6 @@ export default class JoinView extends mixins(validationMixin) {
   box-shadow: var(--box-shadow);
 }
 
-.layout-steps {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-
-  @media (max-width: $breakpoint-m) {
-    display: block;
-    margin: 0;
-  }
-}
-
 .step-title {
   font-size: 1.5rem;
   margin-top: 1rem;
@@ -1138,12 +461,6 @@ export default class JoinView extends mixins(validationMixin) {
   }
 }
 
-.link {
-  font-family: Inter;
-  font-size: 16px;
-  text-decoration: underline;
-}
-
 .cancel-link {
   position: sticky;
   top: 0px;
@@ -1153,23 +470,6 @@ export default class JoinView extends mixins(validationMixin) {
 
 .inputs {
   margin: 1.5rem 0;
-}
-
-.form-background {
-  border-radius: 0.5rem;
-  padding: 1rem;
-  background: var(--bg-light-color);
-  margin-top: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  &:last-of-type {
-    margin-bottom: 2rem;
-  }
-  @media (max-width: $breakpoint-m) {
-    &:first-of-type {
-      margin-top: 0;
-    }
-  }
 }
 
 .input {
@@ -1183,6 +483,7 @@ export default class JoinView extends mixins(validationMixin) {
   font-weight: 400;
   line-height: 24px;
   letter-spacing: 0em;
+  width: 100%;
   &:valid {
     border: 2px solid $clr-green;
   }
@@ -1199,318 +500,5 @@ export default class JoinView extends mixins(validationMixin) {
 
 .input.invalid {
   border: 2px solid var(--error-color);
-}
-
-.input-description {
-  margin-top: 0.25rem;
-  font-size: 14px;
-  font-family: Inter;
-  margin-bottom: 0.5rem;
-  line-height: 150%;
-}
-
-.input-notice {
-  margin-top: 0.25rem;
-  font-size: 12px;
-  font-family: Inter;
-  margin-bottom: 0.5rem;
-  line-height: 150%;
-  color: var(--attention-color);
-  text-transform: uppercase;
-  font-weight: 500;
-}
-
-.input-label {
-  font-family: Inter;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 24px;
-  letter-spacing: 0em;
-  text-align: left;
-  margin: 0;
-}
-
-.radio-row {
-  display: flex;
-  margin-top: 1rem;
-  box-sizing: border-box;
-  border: 2px solid $button-color;
-  border-radius: 1rem;
-  overflow: hidden;
-  width: fit-content;
-  input {
-    position: fixed;
-    opacity: 0;
-    pointer-events: none;
-  }
-  input[type='radio']:checked + label {
-    background: $clr-pink;
-    font-weight: 600;
-  }
-  @media (max-width: $breakpoint-m) {
-    width: 100%;
-    flex-direction: column;
-    text-align: center;
-  }
-}
-
-.radio-btn {
-  box-sizing: border-box;
-  color: white;
-  font-size: 16px;
-  line-height: 24px;
-  align-items: center;
-  padding: 0.5rem 1rem;
-  margin-left: -1px;
-
-  border-right: 2px solid $button-color;
-  border-bottom: none;
-  @media (max-width: $breakpoint-m) {
-    border-right: none;
-    border-bottom: 2px solid $button-color;
-  }
-  &:last-of-type {
-    border-right: none;
-    border-bottom: none;
-  }
-
-  &:hover {
-    opacity: 0.8;
-    background: var(--bg-secondary-color);
-    transform: scale(1.04);
-    cursor: pointer;
-  }
-  &:active {
-    background: var(--bg-secondary-color);
-  }
-}
-
-.loader {
-  display: block;
-  height: 40px;
-  margin: $content-space auto;
-  width: 40px;
-}
-
-.loader:after {
-  content: ' ';
-  display: block;
-  width: 32px;
-  height: 32px;
-  margin: 4px;
-  border-radius: 50%;
-  border: 6px solid #fff;
-  border-color: #fff transparent #fff transparent;
-  animation: loader 1.2s linear infinite;
-}
-
-.loader {
-  margin: $modal-space auto;
-}
-
-@keyframes loader {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.read-only-title {
-  line-height: 150%;
-  margin: 0;
-}
-
-.project-details {
-  &:last-child {
-    margin-bottom: 0;
-  }
-}
-
-.summary {
-  margin-bottom: 1rem;
-}
-
-.summary-section-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 1.5rem;
-  border-bottom: 1px solid $highlight-color;
-  padding-bottom: 0.5rem;
-}
-
-.toggle-tabs-desktop {
-  display: flex;
-  gap: 2rem;
-  font-family: 'Inter';
-  @media (max-width: $breakpoint-m) {
-    /* flex-direction: column;
-    gap: 0;
-    margin-left: 0rem; */
-    /* display: none; */
-  }
-  .active-tab {
-    padding-bottom: 0.5rem;
-    border-bottom: 4px solid $clr-green;
-    border-radius: 4px;
-    font-weight: 600;
-    /* text-decoration: underline; */
-  }
-  .inactive-tab {
-    padding-bottom: 0.5rem;
-    cursor: pointer;
-    &:hover {
-      opacity: 0.8;
-      border-bottom: 4px solid #fff7;
-      border-radius: 4px;
-    }
-    /* text-decoration: underline; */
-  }
-}
-
-.toggle-tabs-mobile {
-  display: flex;
-  gap: 2rem;
-  @media (min-width: $breakpoint-m) {
-    /* flex-direction: column;
-    gap: 0;
-    margin-left: 0rem; */
-    display: none;
-  }
-  .active-tab {
-    padding-bottom: 0.5rem;
-    border-bottom: 4px solid $clr-green;
-    border-radius: 4px;
-    font-weight: 600;
-    /* text-decoration: underline; */
-  }
-  .inactive-tab {
-    padding-bottom: 0.5rem;
-    cursor: pointer;
-    &:hover {
-      opacity: 0.8;
-      transform: scale(1.02);
-    }
-    /* text-decoration: underline; */
-  }
-}
-
-.step-subtitle {
-  margin: 0.5rem 0;
-  font-family: 'Glacial Indifference', sans-serif;
-  font-size: 1.5rem;
-}
-
-.edit-button {
-  font-family: 'Inter';
-  font-weight: 500;
-  font-size: 16px;
-  color: $clr-green;
-}
-
-.data {
-  opacity: 0.8;
-  margin-top: 0.25rem;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.data img {
-  padding: 0.25rem;
-  margin-top: 0.25rem;
-  &:hover {
-    background: var(--bg-primary-color);
-    border-radius: 4px;
-  }
-}
-
-.pt-1 {
-  padding-top: 1rem;
-}
-
-.tx-row {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 0.5rem 0;
-  font-weight: 500;
-}
-
-.tx-row-total {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 0.5rem 0;
-  font-weight: 600;
-  font-size: 20px;
-  font-family: 'Glacial Indifference', sans-serif;
-  background: var(--bg-secondary-color);
-  border-radius: 1rem;
-  padding: 1rem;
-  margin-top: 1.5rem;
-}
-
-.tx-item {
-  font-size: 14px;
-  font-family: Inter;
-  line-height: 150%;
-  text-transform: uppercase;
-  font-weight: 500;
-}
-
-/* .hr {
-  opacity: 0.1;
-  height: 1px;
-  margin: 1rem 0;
-} */
-
-.tx-value {
-}
-
-.fiat-value {
-  font-weight: 400;
-  opacity: 0.8;
-}
-
-.tx-progress-area {
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  margin-bottom: 2rem;
-}
-
-.checkout {
-  margin-bottom: 1rem;
-}
-
-.checkout-desktop {
-  padding-top: 1.5rem;
-  border-top: 1px solid $button-color;
-}
-
-.tx-notice {
-  margin-top: 0.25rem;
-  font-size: 12px;
-  font-family: Inter;
-  margin-bottom: 0.5rem;
-  line-height: 150%;
-  text-transform: uppercase;
-  font-weight: 500;
-}
-
-.break-all {
-  @media (max-width: $breakpoint-s) {
-    display: block;
-  }
-
-  p {
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
 }
 </style>
